Fixes
Navbar bug fixed with brandname component when logged out

refresh token function working on profile page => bug: system looks for non existing user when not logged in crashing app before redirect

updated contect and added comments while trying to troubleshoot authentifiecation redirect issue.

added more to places.name functionality in modal. need to figure out how to add name from one input and combine it with the information from the address input in the modal to send to the user.places DB

redid vehilce and places loop for profile page for ability to remove arrays from the DB then update the compoent. Currently it just needs a route written on the back end for functionality and it needs a way to see the user_id and places/vehicle_id seperatly. Currently things the user_id is the vehicle/places_id (see comments)